amsearch/requirements.txt
Flask==2.3.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
gunicorn==21.2.0
joblib==1.3.0
pgvector==0.2.0
psycopg2==2.9.6
python-dotenv==1.0.0
requests==2.31.0
scikit-learn==1.3.0
sentence-transformers==2.2.2
Werkzeug==2.3.6
